refactor(auth): Movendo os arquivos de tipagem para a pasta auth/types
