+++ src/auth/types/service/payloads/verify-token-payload.type.ts
